Add example: write a 32-bit float value register
I propose to add this example. I've been trying to find information on how to write a multi-register float value on the forum and in the support docs but was unable to find valid information. It makes sense to add this example IMHO to save other people the time and effort to determine how to do this.