Added names in example
It was not obvious to me that the sensor would be called "sensor.stats" and not contain anything from the entity_id field. Adding a name in the example config would probably make it easier for people to find the sensor when they set this up for the first time.