Adding link to releases page for other Pi images
Reading the docs, it seems like the image is only made for the Pi 3 & Pi 4.  This update adds the verbiage to point a user looking for other Pi hardware to the releases page as there is no other place I've found this documentation unless you stumble upon it in releases.