Fixed contradiction for multiple entity IDs
As the `State trigger` section says the following 

> Fires when the state of any of given entities changes.

It makes more sense to me to say if `trigger is true for any entity listed`. As it was originally said it seems to imply that the trigger is only fired when all entities match the condition and not when any of them change as is the case. This also partly fixes #17766 by not having the contradiction in both sections.