Change order of nest instructions in fewer steps
Change the order of nest instructions to reduce the number of times the user has to jump between the device access
console and the cloud console. We start now start with configuring the APIs, configuring OAuth, then turning on device
access and pub/sub.

In follow up changes I am going to change to use *Installed App* credentials so that users no longer need to configure
SSL and URLs, and may automate subscriber configuration.