Correct the documentation about the language
Due to the issue https://github.com/home-assistant/core/issues/57328 it was made clear by @rolfberkenbosch that the described language configuration was in fact erronous. Instead of a two letter language code, the ISO code should have been used. The example was already correct.
I actually took an educated guess as to the default value of the language (en-US). If this is incorrect please let me know.