Update ifttt.markdown
I struggled through setting a simple recipe to receive events from IFTTT. It was not clear to me if I was using the Nabu Casa URL correctly since it did not work for me until  I changed the method to POST. I also had a brief issue understanding the 'event_data action' and so I added a comment to explicitly relate it to the example above. I also reformulated the automation to how I see other website examples (e.g. having Platform first, etc.). Feel free to modify it further if it's not to your likely. Hope this helps others. Thanks.