Added link to Automation Trigger Variables
I can never seen to 'find' the trigger variables (templates) when I look for them. I usually search in the 'Automation Triggers' section looking for them and not remembering that they are on a separate page under Templates, so I just added a link to get there more easily if there are others like me looking for them often.