Update xiaomi_miio.markdown
As a new Roborock S7 vacuum owner, setting up this integration was not straight forward. The Roborock instructions guide you through setting up the device using the Roborock app, but to use this integration, the device must be set up using the Mi Home app as explained on the documents of the underlying python-miio library this integration uses (https://python-miio.readthedocs.io/en/latest/troubleshooting.html). Additionally, having the Roborock S7 on a different subnet or VLAN from Home Assistant does not work without setting up IP masquerading which is also mentioned in the python-miio docs. I added a simple note which highlights this to hopefully help future users.