Add a note in nest about Home Assistant versions
Clarify for users that the documentation on the home assistant website is always for the latest version of Home
Assistant. While this is true of every page on the website, this is a one-off for nest given the setup instructions have
chagned. Users just assume the documentation is out of date rather than realizing their home assistant version is out of
date.