Update controllers.markdown
These sticks are at the top of recommended sticks here, but many people are reporting stability problems with the sticks. The manufacturer, Silabs is apparently aware, but a fix has not been found for about half a year now (see the github page linked). It does not seem like a good idea to recommend them yet.  (Disclaimer: I have one of these sticks, but have not yet implemented it because of these problems.)