Update documentation for nest Media Source attachments
Update documentation for the recent improvements to the nest Media Source and explain how the APIs can be used for
attachments with examples for iOS and Android. The examples are useful because you have to know the different
combinations of camera and phone and how they work together, and they should be useful examples for other contexts too
beyond notifications.