Update RFXtrx logging snippet
When I used the existing snippet, debug got turned on for everything. Setting the default to warning (the default if there is not logging integration) gave me the expected output.

Also, worth turning on debug for any messages that may come from the component (even though there are very few).