Update google.markdown with Troubleshoot 403 error
I ran into a problem of autorization because I hadn't add a test user. Step 11 is not automatically shown as assumed in the walk-thru. So you might skip steps that are necessary  to make it work. Once you run into a problem it is nice that a fix like this is in the troubleshoot section rather than to figure out what step (of 15) you might have missed.