Mention Ecobee developer password 16-character limit
When I was trying to log in to become a developer, when I used a password longer than 16 characters, I noticed it would display as 16 instead, and when I clicked login, it would give no message but not work. There might also (or instead?) be some problems around Firefox as a browser, or possibly my extensions, because the login did work in Edgium, but I wanted to give future folks a heads up.