Added blurb to point users time templating
I had a hard time finding/remembering the `today_at()` variable and was searching here and there with little luck. After finding it where it was not located, I figured adding this blurb for more info at the bottom of the Time and Date might help others find (or discover) that it exists.