Mention short-circuiting issue in `iif`
`iif` filter does not short-circuit which means it is not a universal substitute for templates of the form `<expr> if <conditional> else <expr>`. It can only be used if evaluating all provided expressions does not cause an error. If not you have to use the ternary style expression since that does short-circuit.