Edit sunrise/sunset condition
Edited to follow style used in other conditions (explanation, then example).
Removed non-functional "when_dark" example because it was confusing and in the wrong place.
Moved offset example to a more logical location, added an additional example. 
Moved warning to the end of the section to be congruent with the style of other sections and pages.