Update daikin.markdown
After setting the region code to 'EU' my Daikin unit was greyed out in the ONECTA app and whenever I tried to reconnect it or add it new in a different account I received the error message "Dieses Gerät kann nicht hinzugefügt werden weil die Region Ihres Gerätes nicht unterstützt wird" ("This unit can't be added because your unit's region is not supported") followed by "Region: EU").
After a long support trail with Daikin support I got them to send me an example result of the http://[address]/common/basic_info command and noticed that their 'reg' parameter was set to "eu" not "EU". So I set mine to "eu" and voila, the ONECTA app could recognize the unit again.
So I don't know if all region codes are supposed to be lower case or just the EU one for AC units but there should be a hint at trying the lower case code. This page contains some of the best information on local control of Daikin devices and it may spare other users the same trouble that I went through.
If you have more units to try and you find out that the region code is supposed to be lower case always, I'd suggest to just change all of them to lower case instead.