Update tank_utility.markdown
Provided clarity on requirement for double quotes on each device ID in the configuration.yaml file.  Also corrected configuration markdown to property denote the devices stanza, and not "unit_of_measurement" which is not part of this integration.  Fixed example YAML to include double quotes for those copying and pasting.