Improved the documentation of the tools to read the prof and the callgrind files
When I tried using `gprof2dot`, I was struggling trying to find the correct format.
There is no `-f cprof`, so I tried `-f prof` (which is the default), but the tool can't understand that file in that format.
I had to resort to the forums to find [someone who shared the command-line parameters](https://community.home-assistant.io/t/sharp-permanent-increases-in-cpu-usage/392870/7) he used. Now I'm trying to make the official documentation slightly better, so people won't have to waste too much time with tooling.