Fix error about unit of average power.
The average of something has the same unit of that something. So the average power also has the unit Watt. In fact, to get the average, you would integrate the power, yielding the total energy in kWh, and then divide that by the total time: kWh / h = kW.

The only instance where you would get a unit of W/h (Watt per hour) I can think of, is when you look at the speed at which power consumption changes, i.e. the differential of the power function.