Elaborate on the port setting for Pioneer AVRs
In the section discussing the port number used by Pioneer AVRs,
add a suggestion for how to configure multiple ports to allow
different tools to connect at the same time without blocking
each other.

Verified on Pioneer SC-2023 but the suggestion is probably valid
for most (or all?) network-capable Pioneer AVRs of the same era.