Update statistics.markdown's change* platform
The current documentation says that the change* sensors perform: "The difference between the oldest and newest measurement", and to me, this is misleading. 
The behavior I'm getting (and the one which also makes more sense) is that HA does the "the newest - the oldest" measurement. 

If I have a temp. sensor with a temp that goes from 18 to 22, then the change = Tnew - Told = 22 - 18 = 4 => the temp. has increased