KNX climate preset mode - clarify correct values
The documentation should indicate what the right value is to set the value when using the service 'climate.set_preset_mode'. The errors in the log-files are not helpful. The only acceptable lowercase values seem to be: 'none', 'comfort', 'away', 'sleep' and 'eco'. It might be helpful to indicate these values in the documentation.

https://github.com/home-assistant/home-assistant.io/issues/24190