Fixed broken link to Queensland Bushfire Alert website
The Queensland Bushfire Alert website has been moved to a new domain, and the current URL does not resolve. Using WaybackMachine on the old URL, I was redirected to the new URL, which I have now updated in the documentation accordingly. As neither the old nor the new endpoint links to a GeoRSS feed, I have changed the link text to reflect this.