Removed dead link to the Thinking Cleaner website
The link to the Thinking Cleaner website currently produces an error. Checking the cached version of the site, it seems that the site is not coming up again. The cached versions read: 

> We are out of stock and there will be no new production. We did try but the problem is that some of the components we use in the Thinking Cleaner are not available anymore or if you can find them (or substitutes), those parts are very expensive. For a new production series we would have to redesign with new components and this is very expensive because of certifications and minimum order quantities.

As the archived website contains helpful information about the integration, I have updated the link to reference the archived version.