Update README.md - delete expired/unpurchased domain
Hello from Hacktoberfest,
The domain you are referring to (http://lu-x.me/) is available to everyone to purchase and can be taken over by an attacker. An attacker could exploit such a situation by registering the expired/unpurchased domain and setting up a phishing page that mimics your project/user and etc.