Improve accessibility for expandable details
Improve accessibility of the expandable details items by converting to
a button and tagging with `aria-expanded` so that screen readers can understand
the semantics of these collapsable/expandable units.

A screen reader will now identify the heading, and add "collapsed button" or "expanded button"
depending on the state of the collapsed area. The styling has been updated so that these
buttons look like they did before.