Update calendar.markdown
Added brief note so that users are aware that they need to test with Calendar events further than 15 minutes away, or they might inadvertently think their code / integration is not working. This happened to me on a Google calendar integration and took me quite a bit of time to 'debug' (there was no bug other than creating events to soon in time 😄). I read someone else on the forum with the same problem.