Correct the state attribute example
The `hvac_modes` attribute of a `climate` entity is a list of possible modes, not the current mode (which is the entity state).  I've changed the example to use an attribute that reflects the current `fan_mode` instead.  Someone in Discord pulled that example directly and was confused why it didn't work.