Installation in VirtualBox -> Deactivate paravirtualization
The [issue #71278](https://github.com/home-assistant/core/issues/71278) documents severe problems with timed automations which occur since at least 2019. I can confirm that this still occurs with recent versions (Home Assistant 2022.12.8, Supervisor 2022.12.1, Operating System 9.4), and that it affects more time-related features, which in sum render home assistant useless for many use cases.

The issue contains a fix, which is just changing a single setting of the VM. This setting can also be applied at installation time, so that the problem is prevented from the start. I can confirm that this setting indeed fixes the automation timing issue and several other timing issues that I experienced.

In the issue, there is no sign of any (current or future) change in the code base to fix the bug inside the home-assistant code. So this PR adds the needed instruction to the documentation.