Add links to the support documentation.
It took me a few seconds to understand fully what this extension was about and how to start using it. By being a bit more explicit about the `D` standing for _dynamic_ and adding links to the Google domains documentation, I believe it makes it more accessible. Ironically the "description" meta-data here was already more explicit about the _dynamic_ part but was not visible on the documentation itself.