Remove unnecessary step from setup instructions
The menu in the Google Cloud console has the items "Credentials" and "OAuth Consent Screen" right next to each other on the same level, so it's not needed to click "Credentials" before "OAuth Consent Screen". This might be a copy-paste error as the click on "Credentials" is needed in step 13 (was 14).