Update alexa.smart_home.markdown
I have removed the section about using haaska (which does not seem to be maintained, last doc update was 4 years ago) and the section about how to create other skills.  The vast majority of users just want this to work, and those that are truly looking to create their own skill will figure out how to do it on their own.  I wasted a load of time here trying to setup haaska at this point, following all their outdated docs and finally getting it to somewhat work but not fully.

Then I came back here and skipped the haaska part and everything worked just fine.  Your doc is good.  Don't confuse people with this section.