Update timer.markdown
Edited page to more explicitly say that timers will survive a HA restart if the restore option is used. I was searching through this page with the word 'restart' and it did bring me to the text that mentions if the timer would be remembered during a HA restart, so I added text to say that better for others who my be looking for the same thing.