removed links - simplified
added wording from your comment on backup location
simplified the specification of restore    - could add link to installation-method,, but KISS?

Andsorry for the "send to review",, just starting to contribute on documentation - and have not used github or other developers tools. 
I've been challenged by peers to contribute, and that "it is easy,, you will get plenty of good guidance" ;)  - so I just took a shot :)
Have now tried to read a bit on tpe PR-process, so hopefully my contributions from now on will follow more the standard process. - and sorry for the inconvenience/extra work I've created for you (on 2 out of 3 PR's so far :)  )
 - I stribe to do better  :)