Clarify Tile card documentation image caption
I've clarified the caption of the tile image in the docs to describe it is not showing 1 but 3 tiles.

I came across this page setting up my instance and spent some time trying to figure out how I could get the tile card to display multiple entities. I was sent on this goose chase by this description. Furthermore, I now realize I was very wrong. (If only I read the docs).