Rewrote ZHA docs info on device support and quirks.
Rewrote ZHA docs info on device support and quirks in order to try to better explain "How to add support for new and unsupported devices".  I think that the most important part missing or needing to be clarified was how to submit a new "device support request" and why "quirks" device handlers are needed.