Update Packages folder - simplify
Replacement/simplification of point I was trying to make in #26320. Not sure what happend to that one, I might have cocked up something so it got stuck;)

have tried to simplify;
- removed the option for !include_dir_merge_named from origninal doc, so now this is all about !include_dir_named
- added info on this way of including keep indentation "as if written in configuration.yaml"

The purpose of this PR is to be able to refer to documentation when urging people to use "packages-folders with the indentation of configuration.yaml" 

Currently I've seen several examples where they have read the documentation and set up with !include_dir_merge_named and having to keep up with different indentation than most examples that can be found in communities etc.