Update ecowitt.markdown
It seems that the Ecowitt App prepends a forward slash to the Path field, resulting in
```
//api/webhook/25dce489a3b813257b22fed49c17xxxx
```
which does not work.  Once I removed the extra forward slash, my device and entities then appeared.

I realize this is not a problem with the integration, but the Ecowitt app instead, but it is unclear that you need to remove the first slash from the api token you are given.