Updated Volvo On Call URL and name
Volvo has rebranded Volvo On Call to Volvo Cars App, and has changed the link structure in the process. The current link was dead, so I have updated it to match the new site structure. Since the Home Assistant Integration is still named Volvo On Call, I have left those instance names as is – except for the link to the service, where I have used the new naming convention and explained the name change as a side note.