Removed dead URL to Plugwise Stretch (#29130)
The Plugwise Strech-URL in the docs leads to a dead page. Plugwise has deprecated this product, so I think the right thing to do is remove the link and reference it in the text. I have added a short explanation to the text saying that the product is no longer for sale.