Fixed broken link to Keba manual
The link to the KeContact P20 / P30 UDP Programmers Guide was broken. I have searched through the Keba.com site, and it no longer appears to be available. I have found a copy of the same document on one of their distributors' website and have linked to this file instead.