Add the setup detail that ViCare is required on the device
I ran into the problem that the setup procedure always resulted in an 'invalid credentials' error.

Reason: I was doing the integration setup on my computer. Only after some time it dawned to me that the vicare:// URI probably needs the ViCare application and hence that I needed to run the setup on my mobile phone.

Once I did that I was in business. Since I lost some time finding this out I think it makes sense to explicitly add this information to the setup procedure.