Remove text about needing to reset for config changes from HomeKit
https://github.com/home-assistant/core/pull/101719 adds support for reloading accessories
when the config changes. It should not be needed to reset them anymore when the config
changes.

HomeKit cache does still cache the name so I left that, as well as the block on resetting
because there is always the case where iCloud does not accept the change and it still
does have to be reset, but this should be the execptional case now.