Wake on LAN: Add docs for using helper button & automation (#29586)
* Wake on LAN: Add docs for using helper button & automation

The current docs explain how to create a switch, A switch defined with the `wake_on_lan` platform will render in the UI with both 'on' and 'off' clickable actions. If the user doesn't intend to use the `turn_off` functionality then using a virtual button & automation will look cleaner and less confusing. This change updates the docs to explain how to do that.

* Update wake_on_lan.markdown

* Style tweak: convert into numbered procedure

* indent YAML to fit into step format

---------

Co-authored-by: c0ffeeca7 <38767475+c0ffeeca7@users.noreply.github.com>