Document OS error "Bad message" unhealthy case
With Supervisor pull request https://github.com/home-assistant/supervisor/pull/4750
we've introduced a new unhealthy condition when the operating system
raises "Bad message" errors on file system operations. This typically
points to a corrupted file system.

According to the (opt-in) Sentry error reports we see ~580 affected
systems in the last 30 days. A disproportional amount of systems are
running HAOS on Raspberry Pi. We assume that this is releated to the
known USB UAS storage problem on Raspberry Pi. Add this hints
accordingnly.