docs(logger): Remove special, surprising level
The `logging.NOTSET` level has special meaning in Python and when used results in confusing behavior. Looking at the previous listing of levels a admin not familiar with Python might reasonably assume setting `homeassistant.components.command_line` to `notset` would get them all possible log messages. In fact, it results in no more log messages and thus less log messages than setting `homeassistant.components.command_line` to `debug`.

Better not to list it as an available level. If it needs to be mentioned at all, it should be a special mention that describes the effect it has.