Update octoprint.markdown to add note
To integrate HA and Octoprint, the pop-up only appears when logged in as the user which you are attempting to register.  This may seem logical, but it stumped me and apparently many others from searching for a solution.  Solution found in this comment: https://github.com/home-assistant/core/issues/70152#issuecomment-1100930191