docs: advise reboot of system if wyoming integrations are not being discovered
As described in https://github.com/home-assistant/home-assistant.io/issues/28770, it is possible that the Wyoming integration is not automatically discovered. In that case, adding it manually results in a pop-up configuration box asking for Host and Port fields, and there is no documentation for this. There are a couple of solutions in that issue, but the only thing that worked for me was rebooting my system. (Home Assistant Operating System 12.1, Core 2024.4.3, Supervisor 2024.04.0, on Raspberry Pi 4B.)