Add explicit setup step for opening HTTP/HTTPS port on reolink devices
The Reolink integration requires HTTP/HTTPS ports to be enabled on devices before the it can be configured in Home Assistant. The current setup instructions do not guide users to do this, it is simply left as a note under troubleshooting.

This commit rephrases the installation instructions to ensure users enable HTTP/HTTPS ports.

Because I only have a Video Doorbell WiFi available to me to test, I do not know if this issue occurs with other doorbells. I have therefore phrased the extra installation step conditional on HTTP/HTTPS not yet being available.

As a Linux-only household, this is causing me some difficulty, and I wish I knew about the dependency on a Windows/Mac client before purchasing the device.  Therefore I have also added a note to the list of models indicating this requirement for this model.