Remove Ruby LSP from devcontainer
This extension is causing the Gemfile to be constantly updated in the devcontainer when it's not running on a platform similar to the one that is currently in the repo (ie on an M1 Mac).

This change then ends up in the PR of inexperienced users that have no idea how to deal with it. 

I suggest that we remove it and let the users that work on the Ruby parts of the website, install it manually.