Update serial.markdown
Added 'not', as my initial change was not specific enough.

I am using socat since more than 10 years on other systems. 

Reason of the PR is that the reference to socat and ser2net in this doc already confused a lot of people and made them waste their time (including myself - 2 long evenings of research and try-and-error, without any result).

My conclusion is the same as a bunch of other people had before: There is not legit / stable way to integrate it into HASS OS, as it requires additional driver installations, which are lost on next reboot.