Link to other home automation hubs with some support for MQTT discovery
The intention here isn't to endorse competitors, but to improve the IoT
ecosystem as a whole. Having these links on this page will help users
of those hubs looking to integrate the projects compatible with MQTT
discovery, which in turn gives those projects a larger user base. It's
preferable to list them here rather than have each one attempt to keep
an exhaustive list of compatible software in their own documentation.