Update cgroup_version.markdown
Made this change because you can change the cgroup version by adding kernel boot parameters. This is how they say to do it in docker documentation as well https://docs.docker.com/engine/containers/runmetrics/#changing-cgroup-version.

I had issues with the installer because of my docker cgroup version and this fixed it. I think this is better to point out first because people may forget what they did and putting a more specific course of action can help save them time finding the right solution. Redoing the installation can solve the problem but should not be the first resort as it takes more time.