Update airthings.markdown to reflect the latency in adding the radon sensor.
I noticed that the radon sensor entity was not created when the device was newly added - the Airthings API itself won't return a value until the sensor actually has a known value.  Since this could be confusing, I added a tip to the documention to try and help folks understand it.  This would only happen if you added a new device to Airthings and immediately added it to Home Assistant.