Clarify E1 Pro Hardware Requirements
Added some clarification to the description for the E1 Pro to make it clear that a 5MP or newer camera is required.
Added a footnote that links to the Reolink documentation on hardware versions (mostly just informative, not really anything specific about the remote access functionality, but still seems worthwhile to include) and clarified the requirements.  

Prior documentation was unclear and could have been interpreted that some older version that supported a recently removed feature was required.