Remove SALUS/Computime OTA from ZHA integration support list
Remove mentioning of SALUS/Computime OTA provider support from ZHA integration list of built-in Zigbee OTA providers.

For reference; Support for SALUS/Computime OTA provider was "temporarly" removed in the upstream zigpy a year ago due to old URL is dead but has not yet been updated with a new UR::

* https://github.com/zigpy/zigpy/pull/1522

"***This completely removes the Salus OTA provider from zigpy, as it no longer works due to Salus replacing their app and taking the links offline. Only the config options are kept, so existing configuration doesn't error out for now. In the future, we might be able to partially revert this if we have the new OTA link (assuming the structure is mostly the same).***"